Move SnapFinder from mania ruleset to osu.Game
